Refactor extension and parser modules
