+++ src/parser/binary.ts
@@ -1,33 +1,4 @@
-export interface FieldDef {
-    name: string;
-    type: string;
-    lengthExpr?: string;
-}
-
-export interface StructDef {
-    name: string;
-    fields: FieldDef[];
-}
-
-export interface EnumDef {
-    name: string;
-    underlying: string;
-    entries: Record<number, string>;
-}
-
-export interface FormatDef {
-    structs: Record<string, StructDef>;
-    enums: Record<string, EnumDef>;
-}
-
-export interface TreeNode {
-    name: string;
-    type: string;
-    offset: number;
-    size: number;
-    value?: number | string | (number | string)[];
-    children?: TreeNode[];
-}
+import { FieldDef, StructDef, EnumDef, FormatDef, TreeNode } from './types';
 
 const builtinSizes: Record<string, number> = {
     char: 1,
@@ -38,54 +9,6 @@ const builtinSizes: Record<string, number> = {
     int32_t: 4,
     uint32_t: 4,
 };
-
-export function parseFormatFile(content: string): FormatDef {
-    const structs: Record<string, StructDef> = {};
-    const enums: Record<string, EnumDef> = {};
-
-    const enumRegex = /enum(?:\s+class)?\s+(\w+)(?:\s*:\s*(\w+))?\s*{([\s\S]*?)}\s*;/g;
-    let m: RegExpExecArray | null;
-    while ((m = enumRegex.exec(content)) !== null) {
-        const name = m[1];
-        const underlying = m[2] || 'int32_t';
-        const body = m[3];
-        const entries: Record<number, string> = {};
-        const itemRegex = /(\w+)(?:\s*=\s*([^,]+))?/g;
-        let i: RegExpExecArray | null;
-        let value = 0;
-        while ((i = itemRegex.exec(body)) !== null) {
-            const ename = i[1];
-            const valStr = i[2];
-            if (valStr) {
-                const parsed = parseInt(valStr.trim(), 0);
-                if (!isNaN(parsed)) {
-                    value = parsed;
-                }
-            }
-            entries[value] = ename;
-            value++;
-        }
-        enums[name] = { name, underlying, entries };
-    }
-
-    const structRegex = /struct\s+(\w+)\s*{([\s\S]*?)}\s*;/g;
-    while ((m = structRegex.exec(content)) !== null) {
-        const name = m[1];
-        const body = m[2];
-        const fields: FieldDef[] = [];
-        const lineRegex = /([a-zA-Z_]\w*)\s+([a-zA-Z_]\w*)(\s*\[(.+?)\])?\s*;/g;
-        let f: RegExpExecArray | null;
-        while ((f = lineRegex.exec(body)) !== null) {
-            const type = f[1];
-            const fname = f[2];
-            const lengthExpr = f[4];
-            fields.push({ name: fname, type, lengthExpr });
-        }
-        structs[name] = { name, fields };
-    }
-
-    return { structs, enums };
-}
 
 export function parseBinary(bytes: Uint8Array, format: FormatDef, rootName = 'Root'): TreeNode {
     const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
@@ -177,34 +100,4 @@ export function parseBinary(bytes: Uint8Array, format: FormatDef, rootName = 'Ro
     }
 
     return parseStruct(rootName, {});
-}
-
-let rowCounter = 0;
-
-function treeRows(node: TreeNode, depth = 0): string {
-    const indent = depth * 20;
-    const id = ++rowCounter;
-    const hasChildren = !!(node.children && node.children.length);
-    const value = hasChildren
-        ? ''
-        : node.value !== undefined
-            ? Array.isArray(node.value)
-                ? '[' + node.value.join(', ') + ']'
-                : String(node.value)
-            : '';
-    const arrow = hasChildren ? `<span class="toggle" data-id="${id}">▾</span>` : '';
-    let html = `<tr data-id="${id}" data-depth="${depth}" data-hide-count="0"><td class="name" style="padding-left:${indent}px">${arrow}${node.name}</td>` +
-        `<td>${value}</td><td>${node.type}</td></tr>`;
-    if (node.children) {
-        for (const c of node.children) {
-            html += treeRows(c, depth + 1);
-        }
-    }
-    return html;
-}
-
-export function treeToHtml(node: TreeNode): string {
-    rowCounter = 0;
-    const rows = treeRows(node);
-    return `<table class="tree-table"><thead><tr><th>Name</th><th>Value</th><th>Type</th></tr></thead><tbody>${rows}</tbody></table>`;
 }
