+++ tsconfig.json
@@ -2,7 +2,7 @@
 	"compilerOptions": {
 		"module": "commonjs",
 		"target": "es2020",
-		"lib": ["es2020"],
+                "lib": ["es2020", "dom"],
 		"outDir": "out",
 		"sourceMap": true,
 		"strict": true,
